Rename document.py to xmlparser.py
